Determine cpus_per_task from `jobs.resources.threads` instead of `jobs.threads`
When submitting to the SLURM queue, snakemake requests only a single CPU, no matter what the values of threads or core are. 
A workaround apparently consists in hardcoding the value for `cpus_per_task` in the config.yaml of the profile: 

set-resources:
  - foo:cpus_per_task=8

However, it is not possible to change the `cpus_per_task` dynamically. For instance, the following does not work:

set-resources:
  - foo:cpus_per_task=int(wildcards.processes)

Taking the `cpus_per_task` from `job.resources.threads` instead of `job.threads` gets around this problem. But I am not familiar at all with the snakemake code, so this could break things elsewhere.